[default-login] Extract rendering of login providers to a separate component (and part) for easy override
