+++ packages/@sanity/default-login/sanity.json
@@ -6,12 +6,16 @@
 
   "parts": [
     {
+      "implements": "part:@sanity/base/login-wrapper",
+      "path": "LoginWrapper.js"
+    },
+    {
       "implements": "part:@sanity/base/login-dialog",
       "path": "LoginDialog.js"
     },
     {
-      "implements": "part:@sanity/base/login-wrapper",
-      "path": "LoginWrapper.js"
+      "implements": "part:@sanity/base/login-dialog-content",
+      "path": "LoginDialogContent.js"
     },
     {
       "implements": "part:@sanity/base/authentication-fetcher",
