+++ packages/@sanity/default-login/src/styles/LoginDialogContent.css
@@ -1,0 +1,97 @@
+@import 'part:@sanity/base/theme/variables-style';
+
+.root {
+  display: block;
+  text-align: center;
+}
+
+.title {
+  text-align: center;
+  font-weight: 100;
+  text-transform: uppercase;
+}
+
+.description {
+  text-align: center;
+  font-size: var(--font-size-small);
+  color: var(--text-color);
+}
+
+.sanityLogo {
+  width: 8rem;
+  margin: 4rem auto;
+}
+
+.branding {
+  margin: 0 auto;
+  margin-bottom: 4rem;
+}
+
+.projectName {
+  composes: heading1 from 'part:@sanity/base/theme/typography/text-blocks-style';
+  font-weight: 100;
+}
+
+.projectNameHidden {
+  composes: projectName;
+  display: none;
+}
+
+.brandLogoContainer {
+  height: 3em;
+  @nest & svg {
+    height: 4em;
+    width: auto;
+    max-width: 70vw;
+  }
+}
+
+.providers {
+  display: block;
+  margin: 0;
+  padding: 0;
+  text-align: center;
+}
+
+.inner {
+  display: block;
+}
+
+.provider {
+  display: inline-block;
+  margin: 1em;
+  padding: 0;
+}
+
+.providerButton {
+  appearance: none;
+  display: block;
+  border: none;
+  background-color: transparent;
+  padding: 1em;
+  outline: none !important;
+  cursor: pointer;
+
+  &:hover {
+    background-color: var(--gray-lighter);
+  }
+
+  &:active {
+    background-color: var(--selected-item-color-hover);
+  }
+}
+
+.providerName {
+  display: block;
+}
+
+.providerLogo {
+  @nest & svg {
+    height: 4em;
+    width: auto;
+
+    @media (--screen-medium) {
+      height: 7em;
+    }
+  }
+}
